Fix bug that was skipping inlined functions!
This commit cherry-picks PR #27 to the release v1.0 branch.

Fix bug that was skipping inlined functions!

We were accidentally skipping registering Function protos for inlined
functions, because we were using the frame.pointer to (attempt to)
uniquely identify function StackFrames, but actually inlined function
StackFrames will actually share the same pointer as the function
they were inlined into, since of course all of them are one single frame
in practice, and the pointer refers to the Instruction Pointer's
position in memory, which will be the same for all the inlined frames
and the inlinee.

This meant that the node for a function with other inlined functions was skipping all the inlined functions.

Instead, we need some unique identifier for the function. I'm using the linfo if it's present, and if not I'm using hash((frame.func, frame.file, frame.line)), because at least this will be correct. But I think it might still be overly-unique, because it's possible that multiple lines within the same function will end up registered as different functions. I'm not sure if that matters though?