Fix bug that was skipping inlined functions!
We were accidentally skipping registering Function protos for inlined
functions, because we were using the `frame.pointer` to (attempt to)
uniquely identify function StackFrames, but actually inlined function
StackFrames will actually _share_ the same `pointer` as the function
they were inlined into, since of course all of them are one single frame
in practice, and the `pointer` refers to the Instruction Pointer's
position in memory, which will be the same for all the inlined frames
and the inlinee.

Instead, we need some unique identifier for the function. For now, we
use the hash(frame), which will be unique for every unique IP + Function
combination.