pprof workaround: convert `()` to something pprof won't delete (`（ ）`)
BUT in the process, I've discovered a nicer workaround: It looks like as
long as you set _both_ a Name and an OriginalName that are
**different**, pprof will print the `Name` exactly as-is, unmangled.

I think the assumption is that if you're setting a custom name, then
you know best.

So i can use that to achieve what we want, by supplying both as separate
string identifiers.