Add `full_signatures::Bool` arg to print frames with full argument types (#31)
* Uses `pprof` workaround: It looks like as
long as you set _both_ a Name and an OriginalName that are
**different**, pprof will print the `Name` exactly as-is, unmangled.

I think the assumption is that if you're setting a custom name, then
you know best.

So i can use that to achieve what we want, by supplying both as separate
string identifiers.

* Add `full_signatures::Bool` arg to print frames' typed args!

Also fixes escaping to handle `()` in function names in general. So
before this commit, a function like `(::MyStruct)()` would have printed
as `<unknown>`, and now will print as either `MyStruct` or
`(::MyStruct)()` depending on the value of `full_signatures`. :)

* Add test for `full_signatures`

* Bump to v1.3.0 for full_signatures::Bool