Record a custom (`[unknown function]`) for empty function names
In the current code, we just emit an empty name, which pprof
automatically renders as `<unknown>`, but you cannot
search/select/filter on <unknown>, so we need to give those a proper
name that users can then interact with.

This just renames these to `[unknown function]` so that users can select
and filter it.